refactor: cria arquivos consts, logic, hook, presentation e container no modulo PanelSelector
