+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/consts.tsx
@@ -1,7 +1,7 @@
 import type { CardDetailsMap } from '@helpers/cardsList'; // Adjust the path to where CardDetailsMap is defined
 import type { CardProps } from "@impulsogov/design-system/dist/molecules/Card/Card";
 import type { AcfDashboardType } from '../../../../common/model';
-import type { ExtendedPanelSelectorWithCardsProps } from "./PanelSelector.presentation";
+import type { ExtendedPanelSelectorWithCardsProps } from "./presentation";
 
 // TODO: Esse arquivo vai ser diferente para cada lista e pode ser melhor alterar para um array
 export const externalCardsDetails: CardDetailsMap = {
@@ -157,7 +157,6 @@ export const header = {
 }
 
 //esse conteudo deve ser dinâmico por lista
-
 export const acfNominalListProps = (
     externalCardsProps: CardProps[],
     listName: AcfDashboardType,
