+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/logic.ts
@@ -1,6 +1,7 @@
 import React from "react";
 import type { AcfDashboardType } from "../../../../common/model";
-import type { ExtendedsubTabsWithChildrenAndChildrenDataProps, Tabs } from "./PanelSelector.presentation";
+import type { ExtendedPanelSelectorWithCardsProps, ExtendedsubTabsWithChildrenAndChildrenDataProps, Tabs } from "./presentation";
+import type { PanelSelectorWithCardsProps } from "@impulsogov/design-system/dist/organisms/PanelSelectorWithCards/PanelSelectorWithCards";
 
 export const subTabChildrenSelector = (
     tabs: Tabs,
@@ -30,4 +31,22 @@ export const subTabChildrenSelector = (
             },
             {} as Record<string, React.ReactNode>,
         );
+};
+
+export const tabsBuilder = (
+    props: ExtendedPanelSelectorWithCardsProps,
+    childrenComponents: Record<string, React.ReactNode>
+): PanelSelectorWithCardsProps["tabs"] => {
+    return Object.fromEntries(
+        Object.entries(props.tabs as Tabs).map(([key, tab]) => [
+            key,
+            {
+                ...tab,
+                subTabs: (tab as Tabs[string]).subTabs.map((subTab) => ({
+                    ...subTab,
+                    child: childrenComponents[subTab.subTabID],
+                })),
+            },
+        ])
+    );
 };
