--- src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/sessionHook.ts
@@ -1,8 +1,0 @@
-import type { Session } from "next-auth";
-
-export const sessionHook = async (
-    user: Session["user"] | undefined,
-    setUser: React.Dispatch<React.SetStateAction<Session["user"] | undefined>>
-) => {
-    setUser(user);
-};
